chore: comprehensive architectural refactoring
- Modularized monolithic components (App.tsx, SFTPBrowser.tsx, Terminal.tsx)
- Extracted state management into custom hooks (useConfig, useTabs, useSystemFonts, useUpdateChecker)
- Centralized TypeScript interfaces in src/types.ts
- Created reusable UI components in layout, views, and modals directories
- Improved terminal interaction with ResizeObserver and better focus management
- Consolidated global styles and removed inline CSS from components
- Fixed production build issues and synchronized type imports
- Added ErrorBoundary for improved application stability
