+++ src/App.css
@@ -209,3 +209,33 @@ body.gruvbox-light ::-webkit-scrollbar-thumb:hover {
   border: 3px solid transparent !important;
   background-clip: content-box !important;
 }
+
+/* SFTP Styles */
+.sftp-row {
+    transition: background-color 0.2s;
+}
+.sftp-row:hover {
+    background: rgba(0,0,0,0.05);
+}
+.sftp-row.selected {
+    background: rgba(200, 30, 81, 0.15) !important;
+}
+
+/* Common Animations */
+.spin {
+    animation: spin 1s linear infinite;
+}
+@keyframes spin {
+    0% { transform: rotate(0deg); }
+    100% { transform: rotate(360deg); }
+}
+
+/* Loading Spinner */
+.loading-spinner {
+    width: 30px;
+    height: 30px;
+    border: 3px solid var(--border-color);
+    border-top: 3px solid #c81e51;
+    border-radius: 50%;
+    animation: spin 1s linear infinite;
+}
